no300

+++ leetcode/no300-longest-increasing-subsequence.cc
@@ -1,3 +1,5 @@
+// https://leetcode-cn.com/problems/longest-increasing-subsequence/
+
 #include <iostream>
 #include <vector>
 
@@ -8,8 +10,8 @@ class Solution {
     int res = 0;
     std::vector<int> dp(nums.size(), 1);
 
-    for (int i = 0; i < nums.size(); ++i) {
-      for (int j = 0; j < i; ++j) {
+    for (size_t i = 0; i < nums.size(); ++i) {
+      for (size_t j = 0; j < i; ++j) {
         if (nums[j] > nums[i]) dp[i] = std::max(dp[i], dp[j] + 1);
       }
       res = std::max(res, dp[i]);
